chore: update `with-service-worker` example
