--- examples/with-service-worker/pages/_app.js
@@ -1,25 +1,0 @@
-import { useEffect } from 'react'
-
-function MyApp({ Component, pageProps }) {
-  useEffect(() => {
-    if ('serviceWorker' in navigator) {
-      window.addEventListener('load', function () {
-        navigator.serviceWorker.register('/sw.js').then(
-          function (registration) {
-            console.log(
-              'Service Worker registration successful with scope: ',
-              registration.scope
-            )
-          },
-          function (err) {
-            console.log('Service Worker registration failed: ', err)
-          }
-        )
-      })
-    }
-  }, [])
-
-  return <Component {...pageProps} />
-}
-
-export default MyApp
