--- examples/with-service-worker/pages/index.js
@@ -1,3 +1,0 @@
-export default function Home() {
-  return <h1>with-service-worker</h1>
-}
